Remove project md-data-s3util
