+++ g11n-ws/modules/md-data-api-s3impl/build.gradle
@@ -36,7 +36,6 @@ repositories {
 dependencies {
     compile project(":md-data-api")
     compile project(":vip-common")
-    compile project(":md-data-s3util")
     compile group: 'org.springframework.boot', name: 'spring-boot'
     compile group: 'com.amazonaws', name: 'aws-java-sdk-s3', version: '1.11.550'
 }
