+++ g11n-ws/modules/md-data-api-s3impl/src/main/java/com/vmware/vip/messages/data/dao/impl/S3MultComponentDaoImpl.java
@@ -6,13 +6,10 @@ package com.vmware.vip.messages.data.dao.impl;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
-
 import com.vmware.vip.common.constants.ConstantsChar;
 import com.vmware.vip.messages.data.dao.api.IMultComponentDao;
 import com.vmware.vip.messages.data.dao.api.IOneComponentDao;
@@ -22,7 +19,6 @@ import com.vmware.vip.messages.data.dao.model.ResultI18Message;
  * this class get the bundle json from bundle file
  */
 @Repository
-@Profile("s3")
 public class S3MultComponentDaoImpl implements IMultComponentDao {
    private static Logger logger = LoggerFactory.getLogger(S3MultComponentDaoImpl.class);
