+++ g11n-ws/modules/md-data-api-s3impl/src/main/java/com/vmware/vip/messages/data/dao/impl/S3OneComponentDaoImpl.java
@@ -7,14 +7,11 @@ package com.vmware.vip.messages.data.dao.impl;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
-
 import com.amazonaws.services.s3.model.PutObjectResult;
 import com.amazonaws.services.s3.model.S3Object;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -23,28 +20,27 @@ import com.vmware.vip.common.constants.ConstantsChar;
 import com.vmware.vip.common.constants.ConstantsFile;
 import com.vmware.vip.common.constants.ConstantsKeys;
 import com.vmware.vip.common.i18n.resourcefile.ResourceFilePathGetter;
+import com.vmware.vip.messages.data.conf.S3Cient;
+import com.vmware.vip.messages.data.conf.S3Config;
 import com.vmware.vip.messages.data.dao.api.IOneComponentDao;
 import com.vmware.vip.messages.data.dao.exception.DataException;
 import com.vmware.vip.messages.data.dao.model.ResultI18Message;
-import com.vmware.vip.util.S3Utils;
-import com.vmware.vip.util.conf.S3Client;
-import com.vmware.vip.util.conf.S3Config;
+import com.vmware.vip.messages.data.util.S3Utils;
 
 /**
  * This java class is used to handle translation bundle file or translation
  */
 @Repository
-@Profile("s3")
 public class S3OneComponentDaoImpl implements IOneComponentDao {
 
-   public static final String S3_L10N_BUNDLES_PATH =
-         "l10n" + ConstantsChar.BACKSLASH + "bundles" + ConstantsChar.BACKSLASH;
-
    @Autowired
-   private S3Client s3Client;
+   private S3Cient s3Client;
 
    @Autowired
    private S3Config config;
+
+   private static final String S3_NOT_EXIST_STR = "S3 File doesn't exist: ";
+   private static final String S3_NOT_EXIST_ERR = "File's name doesn't exist!";
 
    private static Logger logger = LoggerFactory.getLogger(S3OneComponentDaoImpl.class);
 
@@ -62,7 +58,7 @@ public class S3OneComponentDaoImpl implements IOneComponentDao {
       } catch (IOException e) {
          String errorLog = ConstantsKeys.FATA_ERROR + e.getMessage();
          logger.error(errorLog, e);
-         throw new DataException(S3Utils.S3_NOT_EXIST_ERR);
+         throw new DataException(S3_NOT_EXIST_ERR);
       }
       if (result != null) {
          result.setProduct(productName);
@@ -70,7 +66,7 @@ public class S3OneComponentDaoImpl implements IOneComponentDao {
          result.setComponent(component);
          result.setLocale(locale);
       } else {
-         throw new DataException(S3Utils.S3_NOT_EXIST_ERR);
+         throw new DataException(S3_NOT_EXIST_ERR);
       }
       return result;
    }
@@ -81,7 +77,7 @@ public class S3OneComponentDaoImpl implements IOneComponentDao {
    @Override
    public String get2JsonStr(String productName, String version, String component, String locale)
          throws DataException {
-      String filePath = S3Utils.genProductVersionS3Path(S3_L10N_BUNDLES_PATH, productName, version) + component
+      String filePath = S3Utils.genProductVersionS3Path(productName, version) + component
             + ConstantsChar.BACKSLASH + ResourceFilePathGetter.getLocalizedJSONFileName(locale);
       String result = null;
       if (s3Client.getS3Client().doesObjectExist(config.getBucketName(), filePath)) {
@@ -91,14 +87,14 @@ public class S3OneComponentDaoImpl implements IOneComponentDao {
                result = S3Utils.convertS3Obj2Str(o);
             } catch (IOException e) {
                logger.warn(e.getMessage(), e);
-               throw new DataException(S3Utils.S3_NOT_EXIST_STR + filePath);
+               throw new DataException(S3_NOT_EXIST_STR + filePath);
             }
          } else {
-            throw new DataException(S3Utils.S3_NOT_EXIST_STR + filePath);
+            throw new DataException(S3_NOT_EXIST_STR + filePath);
          }
       }
       if (result == null) {
-         throw new DataException(S3Utils.S3_NOT_EXIST_STR + filePath);
+         throw new DataException(S3_NOT_EXIST_STR + filePath);
       }
       return result;
    }
@@ -118,7 +114,7 @@ public class S3OneComponentDaoImpl implements IOneComponentDao {
       if (StringUtils.isEmpty(component)) {
          component = ConstantsFile.DEFAULT_COMPONENT;
       }
-      String filePath = S3Utils.genProductVersionS3Path(S3_L10N_BUNDLES_PATH, productName, version) + component
+      String filePath = S3Utils.genProductVersionS3Path(productName, version) + component
             + ConstantsChar.BACKSLASH + ResourceFilePathGetter.getLocalizedJSONFileName(locale);
       Map<String, Object> json = new HashMap<String, Object>();
       json.put(ConstantsKeys.COMPONENT, component);
