--- g11n-ws/modules/md-data-s3util/src/main/java/com/vmware/vip/util/conf/S3Client.java
@@ -1,55 +1,0 @@
-/**
- * Copyright 2019-2020 VMware, Inc.
- * SPDX-License-Identifier: EPL-2.0
- */
-package com.vmware.vip.util.conf;
-
-import javax.annotation.PostConstruct;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.s3.model.GetBucketLocationRequest;
-
-/**
- * the class uses S3 configurations to initialize S3 client environment
- */
-@Configuration
-@Profile("s3")
-public class S3Client {
-    private static Logger logger = LoggerFactory.getLogger(S3Client.class);
-
-    @Autowired
-    private S3Config config;
-
-    private AmazonS3 s3ClientInst;
-
-    /**
-     * initialize the the S3 client environment
-     */
-    @PostConstruct
-    private void init() {
-        s3ClientInst = AmazonS3ClientBuilder.standard()
-                .withCredentials(new AWSStaticCredentialsProvider(
-                        new BasicAWSCredentials(config.getAccessKey(), config.getSecretkey())))
-                .withRegion(config.getS3Region()).enablePathStyleAccess().build();
-        if (!s3ClientInst.doesBucketExistV2(config.getBucketName())) {
-            s3ClientInst.createBucket(config.getBucketName());
-            // Verify that the bucket was created by retrieving it and checking its location.
-            String bucketLocation =
-                    s3ClientInst.getBucketLocation(new GetBucketLocationRequest(config.getBucketName()));
-            logger.info("Bucket location: {}", bucketLocation);
-        }
-    }
-
-    public AmazonS3 getS3Client() {
-        return s3ClientInst;
-    }
-}
