+++ g11n-ws/settings.gradle
@@ -8,7 +8,6 @@ include ':vip-manager-l10n-agent'
 include ':md-data-api'
 include ':md-data-api-pgimpl'
 include ':md-data-initdb'
-include ':md-data-s3util'
 include ':md-data-api-bundleimpl'
 include ':md-restful-api'
 include ':md-restful-api-i18n'
@@ -41,7 +40,6 @@ project(':md-data-api-mt').projectDir         = "$rootDir/modules/md-data-api-mt
 project(':md-data-api-bundleimpl').projectDir = "$rootDir/modules/md-data-api-bundleimpl" as File
 project(':md-data-api-pgimpl').projectDir     = "$rootDir/modules/md-data-api-pgimpl" as File
 project(':md-data-initdb').projectDir         = "$rootDir/modules/md-data-initdb" as File
-project(':md-data-s3util').projectDir         = "$rootDir/modules/md-data-s3util" as File
 project(':md-data-api').projectDir            = "$rootDir/modules/md-data-api" as File
 project(':tool-trans-fetcher').projectDir     = "$rootDir/tools/tool-trans-fetcher" as File
 project(':vip-resources').projectDir          = "$rootDir/vip-resources" as File
