+++ g11n-ws/vip-manager-l10n/build.gradle
@@ -59,7 +59,6 @@ dependencies {
     compile("io.jsonwebtoken:jjwt:0.9.0")
 
     compile group: 'com.amazonaws', name: 'aws-java-sdk-s3', version: '1.11.550'
-    compile project(":md-data-s3util")
     testCompile("org.springframework.boot:spring-boot-starter-test")
 }
